+++ pkg/microservice/aslan/core/environment/handler/istio_grayscale.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 
 	"github.com/gin-gonic/gin"
+	"github.com/koderover/zadig/v2/pkg/microservice/aslan/core/common/service/kube"
 	"github.com/koderover/zadig/v2/pkg/microservice/aslan/core/environment/service"
 	"github.com/koderover/zadig/v2/pkg/setting"
 	internalhandler "github.com/koderover/zadig/v2/pkg/shared/handler"
@@ -246,7 +247,7 @@ func GetIstioGrayscaleConfig(c *gin.Context) {
 // @Produce json
 // @Param 	projectName	query		string									true	"project name"
 // @Param 	name 		path		string									true	"env name"
-// @Param 	body 		body 		service.SetIstioGrayscaleConfigRequest 	true 	"body"
+// @Param 	body 		body 		kube.SetIstioGrayscaleConfigRequest 	true 	"body"
 // @Success 200
 // @Router /api/aslan/environment/production/environments/{name}/istioGrayscale/config [post]
 func SetIstioGrayscaleConfig(c *gin.Context) {
@@ -282,7 +283,7 @@ func SetIstioGrayscaleConfig(c *gin.Context) {
 		}
 	}
 
-	req := service.SetIstioGrayscaleConfigRequest{}
+	req := kube.SetIstioGrayscaleConfigRequest{}
 	err = c.ShouldBindJSON(&req)
 	if err != nil {
 		ctx.Err = e.ErrInvalidParam.AddErr(err)
