+++ pkg/setting/consts.go
@@ -885,3 +885,12 @@ const (
 const (
 	WorkflowTimeFormat = "[2006-01-02 15:04:05]"
 )
+
+const (
+	IstioNamespace                          = "istio-system"
+	IstioProxyName                          = "istio-proxy"
+	ZadigEnvoyFilter                        = "zadig-share-env"
+	EnvoyFilterNetworkHttpConnectionManager = "envoy.filters.network.http_connection_manager"
+	EnvoyFilterHttpRouter                   = "envoy.filters.http.router"
+	EnvoyFilterLua                          = "type.googleapis.com/envoy.extensions.filters.http.lua.v3.Lua"
+)
